pyproject.toml
[project]
name = "bill-retrieval-chatbot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy>=2.0",
    "alembic",
    "psycopg2-binary",
    "lxml",
    "loguru",
    "typer",
    "pydantic-settings",
    "sentence-transformers",
    "openai",
    "anthropic",
    "pgvector>=0.4.2",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "httpx", "lxml-stubs", "pytest-playwright"]

[tool.pytest.ini_options]
testpaths = ["tests"]
